prevent unnecessary buffer copying
hydrogen_mp_obj_bytes replaces mp_obj_new_bytes and uses the buffer that is
passed in directly as the buffer for the bytes object.

this reduces the peak memory usage by half, this is very significant for
functions like hydrogen_random_buf that used twice the requested memory while
copying the buffer in mp_obj_new_bytes.